update toolbar, drawer - add isLogged
add AuthNav, UserMenu
