--- src/components/Drawer/AppBarNotLogged.jsx
@@ -1,69 +1,0 @@
-import AppBar from '@mui/material/AppBar';
-import Box from '@mui/material/Box';
-import Toolbar from '@mui/material/Toolbar';
-import Typography from '@mui/material/Typography';
-
-function AppBarNotLogged() {
-  return (
-    <AppBar
-      position="fixed"
-      // sx={{ zIndex: theme => theme.zIndex.drawer + 1 }}
-    >
-      <Toolbar>
-        <Typography
-          variant="h6"
-          noWrap
-          component="a"
-          href="/"
-          sx={{
-            mr: 10,
-            display: { xs: 'none', md: 'flex' },
-            fontFamily: 'Roboto',
-            fontWeight: 700,
-            letterSpacing: '.3rem',
-            color: 'inherit',
-            textDecoration: 'none',
-          }}
-        >
-          FIFO
-        </Typography>
-        <Typography
-          variant="h5"
-          noWrap
-          component="a"
-          href="/"
-          sx={{
-            mr: 2,
-            display: { xs: 'flex', md: 'none' },
-            flexGrow: 1,
-            fontFamily: 'Roboto',
-            fontWeight: 700,
-            letterSpacing: '.3rem',
-            color: 'inherit',
-            textDecoration: 'none',
-          }}
-        >
-          FIFO
-        </Typography>
-        <Box sx={{ flexGrow: 1, display: { xs: 'none', md: 'flex' } }}>
-          <Typography
-            variant="body1"
-            noWrap
-            component="p"
-            sx={{
-              mr: 2,
-              display: { xs: 'none', md: 'flex' },
-              fontFamily: 'monospace',
-              fontWeight: 500,
-              letterSpacing: '.15rem',
-              color: 'inherit',
-            }}
-          >
-            First in - First out
-          </Typography>
-        </Box>
-      </Toolbar>
-    </AppBar>
-  );
-}
-export default AppBarNotLogged;
