+++ src/components/Drawer/DrawerMUI.jsx
@@ -23,9 +23,10 @@ import { drawerWidth } from 'constants/drawer';
 import { NavLink as RouterNavLink } from 'react-router-dom';
 import ListItemText from '@mui/material/ListItemText';
 
+const isLoggedIn = true;
 
 // ['Сформувати звіт', 'Корисні відео', 'Нормативи']
-const arrLinks = ['/generate-report','/usefull-material','/normative' ]
+const arrLinks = ['/generate-report', '/usefull-material', '/normative'];
 
 function DrawerMUI(props) {
   const { window, children } = props;
@@ -39,40 +40,38 @@ function DrawerMUI(props) {
     <div>
       <Toolbar />
       <Toolbar />
-<nav>
-      <Broker
-        nestedList={{
-          primaryTitle: 'Брокер',
-          nestedItems: ['Freedom Finance', 'Interactive Brokers'],
-          nestedLinks: ['/freedom-finance', '/interactive-brokers']
-        }}
-      />
+      <nav>
+        <Broker
+          nestedList={{
+            primaryTitle: 'Брокер',
+            nestedItems: ['Freedom Finance', 'Interactive Brokers'],
+            nestedLinks: ['/freedom-finance', '/interactive-brokers'],
+          }}
+        />
 
-      <Operations
-        nestedList={{
-          primaryTitle: 'Операції',
-          nestedItems: ['Купити', 'Продати'],
-          nestedLinks: ['/buy', '/sell']
+        <Operations
+          nestedList={{
+            primaryTitle: 'Операції',
+            nestedItems: ['Купити', 'Продати'],
+            nestedLinks: ['/buy', '/sell'],
+          }}
+        />
+        <Divider />
+        <List>
+          {['Сформувати звіт', 'Корисні відео', 'Нормативи'].map(
+            (text, index) => (
+              <ListItem key={text} disablePadding>
+                <ListItemButton component={RouterNavLink} to={arrLinks[index]}>
+                  <ListItemIcon>
+                    {index % 2 === 0 ? <InboxIcon /> : <MailIcon />}
+                  </ListItemIcon>
 
-        }}
-      />
-      <Divider />
-      <List>
-        {['Сформувати звіт', 'Корисні відео', 'Нормативи'].map(
-          (text, index) => (
-            <ListItem key={text} disablePadding>
-              <ListItemButton component={RouterNavLink}
-                        to={arrLinks[index]}>
-                <ListItemIcon>
-                  {index % 2 === 0 ? <InboxIcon /> : <MailIcon />}
-                </ListItemIcon>
-                 
-                        <ListItemText primary={text} />
-              </ListItemButton>
-            </ListItem>
-          )
-        )}
-      </List>
+                  <ListItemText primary={text} />
+                </ListItemButton>
+              </ListItem>
+            )
+          )}
+        </List>
       </nav>
     </div>
   );
@@ -81,7 +80,7 @@ function DrawerMUI(props) {
     window !== undefined ? () => window().document.body : undefined;
 
   return (
-    <Box sx={{ display: 'flex' }}>
+    <Box sx={{ display: 'flex', justifyContent: 'center' }}>
       <CssBaseline />
       <AppBar
         position="fixed"
@@ -98,72 +97,84 @@ function DrawerMUI(props) {
             <MenuIcon />
           </IconButton>
 
-          <ContentToolbarForDrawer />
+          <ContentToolbarForDrawer isLoggedIn={isLoggedIn} />
         </Toolbar>
       </AppBar>
-      <Box
-        component="nav"
-        sx={{ width: { lg: drawerWidth }, flexShrink: { sm: 0 } }}
-        aria-label="mailbox folders"
-      >
-        {/* The implementation can be swapped with js to avoid SEO duplication of links. */}
-        <Drawer
-          container={container}
-          variant="temporary"
-          open={mobileOpen}
-          onClose={handleDrawerToggle}
-          ModalProps={{
-            keepMounted: true, // Better open performance on mobile.
-          }}
-          sx={{
-            display: { xs: 'block', lg: 'none' },
-            '& .MuiDrawer-paper': {
-              boxSizing: 'border-box',
-              width: drawerWidth,
-            },
-          }}
-        >
-          {drawer}
-        </Drawer>
-        <Drawer
-          variant="permanent"
-          sx={{
-            display: { xs: 'none', lg: 'block' },
-            '& .MuiDrawer-paper': {
-              boxSizing: 'border-box',
-              width: drawerWidth,
-            },
-          }}
-          open
-        >
-          {drawer}
-        </Drawer>
-      </Box>
-      <Box
-        component="main"
-        sx={{
-          flexGrow: 1,
-          p: 3,
-          width: { lg: `calc(100% - ${drawerWidth}px)` },
-        }}
-      >
-        <Toolbar />
-        <Box
-          component="div"
-          sx={{
-            flexGrow: 1,
-            p: 3,
-            width: { lg: `calc(100% - ${drawerWidth}px - 32 px)` },
-            height: '80vh',
-            background: '#ddcece',
-            borderRadius: 2,
-            overflow: 'auto',
-          }}
-        >
-          <BrokerInformation />
-          {children}
-        </Box>
-      </Box>
+
+      {isLoggedIn && (
+        <>
+          <Box
+            component="nav"
+            sx={{ width: { lg: drawerWidth }, flexShrink: { sm: 0 } }}
+            aria-label="mailbox folders"
+          >
+            {/* The implementation can be swapped with js to avoid SEO duplication of links. */}
+            <Drawer
+              container={container}
+              variant="temporary"
+              open={mobileOpen}
+              onClose={handleDrawerToggle}
+              ModalProps={{
+                keepMounted: true, // Better open performance on mobile.
+              }}
+              sx={{
+                display: { xs: 'block', lg: 'none' },
+                '& .MuiDrawer-paper': {
+                  boxSizing: 'border-box',
+                  width: drawerWidth,
+                },
+              }}
+            >
+              {drawer}
+            </Drawer>
+            <Drawer
+              variant="permanent"
+              sx={{
+                display: { xs: 'none', lg: 'block' },
+                '& .MuiDrawer-paper': {
+                  boxSizing: 'border-box',
+                  width: drawerWidth,
+                },
+              }}
+              open
+            >
+              {drawer}
+            </Drawer>
+          </Box>
+        </>
+      )}
+
+      {isLoggedIn ? (
+        <>
+          <Box
+            component="main"
+            sx={{
+              flexGrow: 1,
+              p: 3,
+              width: { lg: `calc(100% - ${drawerWidth}px)` },
+            }}
+          >
+            <Toolbar />
+            <Box
+              component="div"
+              sx={{
+                flexGrow: 1,
+                p: 3,
+                width: { lg: `calc(100% - ${drawerWidth}px - 32 px)` },
+                height: '80vh',
+                background: '#ddcece',
+                borderRadius: 2,
+                overflow: 'auto',
+              }}
+            >
+              <BrokerInformation />
+              {children}
+            </Box>
+          </Box>
+        </>
+      ) : (
+        <>{children}</>
+      )}
     </Box>
   );
 }
