+++ src/components/Drawer/ToolbarForDrawer.jsx
@@ -1,47 +1,12 @@
 import { useState } from 'react';
 import Box from '@mui/material/Box';
-import IconButton from '@mui/material/IconButton';
 import Typography from '@mui/material/Typography';
-import Menu from '@mui/material/Menu';
-import Avatar from '@mui/material/Avatar';
-import Tooltip from '@mui/material/Tooltip';
-import MenuItem from '@mui/material/MenuItem';
-import LogoutIcon from '@mui/icons-material/Logout';
 import { Link as RouterLink } from 'react-router-dom';
+import UserMenu from './UserMenu';
+import AuthNav from './AuthNav';
+import { Link } from '@mui/material';
 
-
-const settings = ['Profile', 'Account', 'Dashboard', 'Choose theme'];
-
-function stringAvatar(name) {
-  return {
-    sx: {
-      bgcolor: stringToColor(name),
-    },
-    children: `${name.split(' ')[0][0]}${name.split(' ')[1][0]}`,
-  };
-}
-
-function stringToColor(string) {
-  let hash = 0;
-  let i;
-
-  /* eslint-disable no-bitwise */
-  for (i = 0; i < string.length; i += 1) {
-    hash = string.charCodeAt(i) + ((hash << 5) - hash);
-  }
-
-  let color = '#';
-
-  for (i = 0; i < 3; i += 1) {
-    const value = (hash >> (i * 8)) & 0xff;
-    color += `00${value.toString(16)}`.slice(-2);
-  }
-  /* eslint-enable no-bitwise */
-
-  return color;
-}
-
-function ContentToolbarForDrawer() {
+function ContentToolbarForDrawer({ isLoggedIn }) {
   const [anchorElUser, setAnchorElUser] = useState(null);
 
   const handleOpenUserMenu = event => {
@@ -54,11 +19,12 @@ function ContentToolbarForDrawer() {
 
   return (
     <>
-      <Typography
+      <Link
         variant="h6"
         noWrap
         component={RouterLink}
-       to="/"
+        to="/"
+        underline="hover"
         sx={{
           mr: 10,
           display: { xs: 'none', md: 'flex' },
@@ -70,12 +36,13 @@ function ContentToolbarForDrawer() {
         }}
       >
         FIFO
-      </Typography>
-      <Typography
+      </Link>
+      <Link
         variant="h5"
         noWrap
         component={RouterLink}
-       to="/"
+        to="/"
+        underline="hover"
         sx={{
           mr: 2,
           display: { xs: 'flex', md: 'none' },
@@ -88,7 +55,7 @@ function ContentToolbarForDrawer() {
         }}
       >
         FIFO
-      </Typography>
+      </Link>
       <Box sx={{ flexGrow: 1, display: { xs: 'none', md: 'flex' } }}>
         <Typography
           variant="body1"
@@ -106,55 +73,16 @@ function ContentToolbarForDrawer() {
           First in - First out
         </Typography>
       </Box>
-      <Box sx={{}}>
-        <Typography
-          variant="body1"
-          noWrap
-          component="p"
-          sx={{
-            mr: 2,
-            display: { xs: 'none', md: 'flex' },
-            fontFamily: 'Roboto',
-            fontWeight: 500,
-            letterSpacing: '.15rem',
-            color: 'inherit',
-          }}
-        >
-          User name
-        </Typography>
-      </Box>
-      <Box sx={{ flexGrow: 0 }}>
-        <Tooltip title="Open settings">
-          <IconButton onClick={handleOpenUserMenu} sx={{ p: 0, mr: 2 }}>
-            <Avatar {...stringAvatar('Kent Dodds')} />
-          </IconButton>
-        </Tooltip>
-        <Menu
-          sx={{ mt: '45px' }}
-          id="menu-appbar"
-          anchorEl={anchorElUser}
-          anchorOrigin={{
-            vertical: 'top',
-            horizontal: 'right',
-          }}
-          keepMounted
-          transformOrigin={{
-            vertical: 'top',
-            horizontal: 'right',
-          }}
-          open={Boolean(anchorElUser)}
-          onClose={handleCloseUserMenu}
-        >
-          {settings.map(setting => (
-            <MenuItem key={setting} onClick={handleCloseUserMenu}>
-              <Typography textAlign="center">{setting}</Typography>
-            </MenuItem>
-          ))}
-        </Menu>
-        <IconButton aria-label="logout">
-          <LogoutIcon sx={{ color: '#f39fad' }} />
-        </IconButton>
-      </Box>
+
+      {isLoggedIn ? (
+        <UserMenu
+          valueAnchorElUser={anchorElUser}
+          onCloseUserMenu={handleCloseUserMenu}
+          onOpenUserMenu={handleOpenUserMenu}
+        />
+      ) : (
+        <AuthNav />
+      )}
     </>
   );
 }
