+++ src/layout/Layout.jsx
@@ -1,17 +1,17 @@
 import { Suspense } from 'react';
 import { Outlet } from 'react-router-dom';
 
-import Box from '@mui/material/Box';
+// import Box from '@mui/material/Box';
 
-import ResponsiveDrawer from 'components/Drawer';
-import AppBarNotLogged from 'components/Drawer/AppBarNotLogged';
+import DrawerMUI from 'components/Drawer';
+// import AppBarNotLogged from 'components/Drawer/AppBarNotLogged';
 import Footer from 'components/Footer';
 
 export default function Layout() {
-  const isLoggedIn = false;
+  // const isLoggedIn = true;
   return (
     <div>
-      {!isLoggedIn && (
+      {/* {!isLoggedIn && (
         <div>
           <AppBarNotLogged />
           <Box sx={{ height: '100px' }}></Box>
@@ -19,15 +19,15 @@ export default function Layout() {
             <Outlet />
           </Suspense>
         </div>
-      )}
+      )} */}
 
-      {isLoggedIn && (
-        <ResponsiveDrawer>
-          <Suspense fallback={null}>
-            <Outlet />
-          </Suspense>
-        </ResponsiveDrawer>
-      )}
+      {/* {isLoggedIn && ( */}
+      <DrawerMUI>
+        <Suspense fallback={null}>
+          <Outlet />
+        </Suspense>
+      </DrawerMUI>
+      {/* )} */}
       <Footer />
     </div>
   );
